dev-requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-celery-results==2.5.0
django-filter==22.1
django-quill-editor==0.1.40
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
kombu==5.2.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.2
pandas==1.5.3
pandas-stubs==1.5.3.230321
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
types-pytz==2022.7.1.2
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
